ようこその編集

+++ app/Http/Controllers/TodoController.php
@@ -10,23 +10,40 @@ class TodoController extends Controller
 {
     public function index(Request $request)
     {
-           //ログインしているユーザーのIDを取得
-            $user_id = Auth::id();
-            //検索キーワードを取得
-            $keyword = $request->input('keyword');
-            $query = Todolist::query();
-            //もし検索キーワードが入力されていれば、検索結果を取得&ログインしているユーザーのタスクを変数に設定
-            //検索キーワードが入力されていなければ、ログインしているユーザーのタスクを変数に設定
-            if(!empty($keyword)) {
-                $query->where('name', 'LIKE', "%{$keyword}%")
-                    ->where('user_id', '=', "$user_id");
-            } else {
-                $query->where('user_id', '=', "$user_id");
-            }
-            //設定した変数の情報を期限が違い順に取得して、indexに返す
-            $todolists = $query->orderByRaw('limit_date')->get();
-            return view('todolist.index', compact('todolists', 'keyword'));
+        //ログインしているユーザーのIDを取得
+        $user_id = Auth::id();
+        //検索キーワードを取得
+        $keyword = $request->input('keyword');
+        //選択されたステータスを取得
+        $status = $request->input('status');
+        //期限切れ判定用
+        $today = date('Y-m-d');
 
+        $query = Todolist::query();
+        //ログインしているユーザーのタスクを表示する
+        $query->where('user_id', '=', $user_id);
+        //もし検索キーワードが入力されていれば、検索結果を取得&ログインしているユーザーのタスクを変数に設定
+        if(!empty($keyword)) {
+            $query->where('name', 'LIKE', "%{$keyword}%");
+        }
+        //もしステータスが選択されていれば、ステータスの絞り込み
+        if(!empty($status)) {
+            if($status == "done") {
+                $query->whereNotNull('completion_date');
+            } elseif($status == "past") {
+                $query->whereNotNull('limit_date')
+                    ->where('limit_date', '<', "$today");
+            } elseif($status == "work") {
+                $query->where(function ($query) use ($today){
+                    $query->whereNull('completion_date')
+                    ->where('limit_date', '>', "$today")
+                    ->orwhereNull('limit_date');
+                });
+            }
+        }
+        //設定した変数の情報を期限が違い順に取得して、indexに返す
+        $todolists = $query->orderByRaw('limit_date')->get();
+        return view('todolist.index', compact('todolists', 'keyword', 'status'));
     }
 
     public function create()
@@ -62,7 +79,7 @@ class TodoController extends Controller
         $todolists = Todolist::find($id);
         //もし、ログインしているユーザーのIDと登録されているユーザーIDが違えばエラー画面に遷移
         if(Auth::id() != $todolists->user_id) {
-            return abort('404');
+            return redirect()->route('todo.login');
         } else {
             return view('todolist.show', [
                 "todolists" => $todolists
@@ -74,7 +91,7 @@ class TodoController extends Controller
     {
         $todolists = Todolist::find($id);
         if(Auth::id() != $todolists->user_id) {
-            return abort('404');
+            return redirect()->route('todo.login');
         } else {
             return view('todolist.edit', [
                 "todolists" => $todolists
@@ -87,13 +104,17 @@ class TodoController extends Controller
         try {
             DB::beginTransaction();
             $todolists = Todolist::find($request->id);
-            $todolists->name  =  $request->name;
-            $todolists->text  =  $request->text;
-            $todolists->limit_date  =  $request->limit_date;
-            $todolists->completion_date  =  $request->completion_date;
-            $todolists->save();
-            DB::commit();
-            return redirect()->route('todo.index');
+            if(Auth::id() != $todolists->user_id) {
+                return redirect()->route('todo.login');
+            } else {
+                $todolists->name  =  $request->name;
+                $todolists->text  =  $request->text;
+                $todolists->limit_date  =  $request->limit_date;
+                $todolists->completion_date  =  $request->completion_date;
+                $todolists->save();
+                DB::commit();
+                return redirect()->route('todo.index');
+            }
         } catch (\Exception $e) {
             DB::rollBack();
         }
@@ -104,8 +125,12 @@ class TodoController extends Controller
 
     public function softDeletes($id)
     {
-        Todolist::find($id)->delete();
-        return redirect()->route('todo.index');
+        $todolists = Todolist::find($id);
+        if(Auth::id() != $todolists->user_id) {
+            return redirect()->route('todo.login');
+        } else {
+            Todolist::find($id)->delete();
+            return redirect()->route('todo.index');
+        }
     }
-
 }
